fix: split atomic_write
